Move services and pipelines under one supervisor
